Delete unnecessary Entity
