--- domain/Entities/Registers/UserNickNameEntity.php
@@ -1,79 +1,0 @@
-<?php
-namespace Domain\Entities\Registers;
-
-use Domain\ValueObjects\NickNameValueObject;
-use Domain\ValueObjects\TimeStampValueObject;
-use Illuminate\Contracts\Support\Arrayable;
-use stdClass;
-
-/**
- * Class RegisterUserNickNameEntity
- * @package Domain\Entities
- */
-class UserNickNameEntity implements Arrayable
-{
-    private $id;
-    private $nickname;
-    private $timeStamp;
-
-    /**
-     * RegisterUserPasswordEntity constructor.
-     * @param int $userId
-     * @param stdClass $userRecord
-     */
-    public function __construct(
-        int $userId,
-        stdClass $userRecord
-    ) {
-        $this->id        = $userId;
-        $this->nickname  = new NickNameValueObject($userRecord);
-        $this->timeStamp = new TimeStampValueObject();
-    }
-
-    /**
-     * Get the instance as an array.
-     *
-     * @return array
-     */
-    public function toArray(): array
-    {
-        return [
-            'id'         => $this->getId(),
-            'nickname'   => $this->getNickName(),
-            'created_at' => $this->getCreatedAt(),
-            'updated_at' => $this->getUpdatedAt(),
-        ];
-    }
-
-    /**
-     * @return int
-     */
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getNickName(): string
-    {
-        return $this->nickname->getNickName();
-    }
-
-    /**
-     * @return string
-     */
-    public function getCreatedAt(): string
-    {
-        return $this->timeStamp->getNow();
-    }
-
-    /**
-     * @return string
-     */
-    public function getUpdatedAt(): string
-    {
-        return $this->timeStamp->getNow();
-    }
-}
